move UpdateFieldsBenchmark tests to ColumnExpressionSuite
